cleanup(fancydesktop): Only copying and renaming if not already done

+++ installer/targets/tue-fancydesktop/install.bash
@@ -3,10 +3,13 @@ source /etc/lsb-release
 
 # copying shortcuts to desktop
 cp $TUE_DIR/installer/targets/tue-fancydesktop/shortcuts/.*.desktop ~/Desktop
-cp $TUE_DIR/installer/targets/tue-fancydesktop/shortcuts/.*.desktop ~/.local/share/applications/
-find ~/.local/share/applications/ -iname ".*.desktop" -exec rename 's/\.Terminator/Terminator/' '{}' \;
-find ~/.local/share/applications/ -iname ".*.desktop" -exec rename 's/\.AMIGO/AMIGO/' '{}' \;
-find ~/.local/share/applications/ -iname ".*.desktop" -exec rename 's/\.SERGIO/SERGIO/' '{}' \;
+if [ ! -f ~/.local/share/applications/AMIGO\ Dashboard.desktop ]
+then
+    cp $TUE_DIR/installer/targets/tue-fancydesktop/shortcuts/.*.desktop ~/.local/share/applications/
+    find ~/.local/share/applications/ -iname ".*.desktop" -exec rename 's/\.Terminator/Terminator/' '{}' \;
+    find ~/.local/share/applications/ -iname ".*.desktop" -exec rename 's/\.AMIGO/AMIGO/' '{}' \;
+    find ~/.local/share/applications/ -iname ".*.desktop" -exec rename 's/\.SERGIO/SERGIO/' '{}' \;
+fi
 
 # copying shortcuts to desktop
 if [ ! -d /usr/share/pixmaps/tue ]
